move service lifecycle test in a separate file
